Implement command runner using job_start()
This allows to capture stderr. system() command would concatenate stderr
and stdout, what is kind a wrong, because sometimes the code contains
diagnostic warnings or errors, and during the on-demand compilation
within the Tsepepe tool, some errors might be put to stderr.
We need to separate it, because otherwise the errors will appear in
stdout, thus it will land in the definition file, along the function
definition.